requirements/common.txt
appdirs==1.4.3
awesome-slugify==1.6.4
Babel==2.4.0
defusedxml==0.5.0
Django==1.11
django-allauth==0.31.0
django-bootstrap-form==3.2.1
django-braces==1.11.0
django-debug-toolbar==1.8
django-extensions==1.7.9
django-phonenumber-field==1.3.0
easy-thumbnails==2.4.1
oauthlib==2.0.2
packaging==16.8
phonenumberslite==8.4.2
Pillow==3.1.1
psycopg2==2.7.1
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
regex==2017.4.29
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.20
